+++ ChangeLog.md
@@ -5,6 +5,7 @@ where provided, and the git commit history.
 
 * 2017/03/?? - 1.67 - PENDING
     * Refactored a number of the view params and added Module as a param (Jazz: 461738)
+    * Fixed Success messages to show in green banner, added success messages to concept create and edit, optimized some of the concept editing code
 
 * 2017/03/08 - 1.66
     * Added role "alert" to error messages so they will be read by screen readers. They do not take the focus so that they don't interrupt the user flow, and because we add multiple at a time. Some messages can't contain detailed descriptions of the item affected if it is a new item. (Jazz: 469380)