+++ ChangeLog.md
@@ -3,7 +3,7 @@ Komet Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers 
 where provided, and the git commit history.
 
-* 2017/03/?? - 1.67 - PENDING
+* 2017/03/15 - 1.67
     * Refactored a number of the view params and added Module as a param (Jazz: 461738)
     * Fixed Success messages to show in green banner, added success messages to concept create and edit, optimized some of the concept editing code
     * Added status preference to taxonomy and concept viewer (Jazz: 461738)